+++ pkg/README.md
@@ -8,11 +8,11 @@ The packages in this directory provide a set of tools that users can use to inte
 
 Deploy the MCMS contract using the [`DeployManyChainMultiSig`](./gethwrappers/ManyChainMultiSig.go#L76) function
 
-The default [`ManyChainMultiSigConfig`](./gethwrappers/ManyChainMultiSig.go#L32) is very unintuitive to define and construct based on the desired group structure. As a result, this library provides a [`Config` Wrapper](./configwrappers/config.go#L13) which defines a more intuitive MCMS membership structure for ease of use.
+The default [`ManyChainMultiSigConfig`](./gethwrappers/ManyChainMultiSig.go#L32) is very unintuitive to define and construct based on the desired group structure. As a result, this library provides a [`Config` Wrapper](config/config.go#L13) which defines a more intuitive MCMS membership structure for ease of use.
 
 The configuration is a nested tree structure where a given group is considered at `quorum` if the sum of `Signers` with Signatures and `GroupSigners` that individually are at `quorum` are greater than or equal to the top-level `quorum`
 
-For example, given the following [`Config`](./configwrappers/config.go#L13):
+For example, given the following [`Config`](config/config.go#L13):
 
 ```
 Config{
@@ -48,7 +48,7 @@ This configuration represents a membership structure that requires 3 entities to
 11. [`0x2`, `0x4`, `0x5`]
 12. [`0x2`, `0x4`, `0x6`]
 
-Once a satisfactory MCMS Membership configuration is constructed, users can use the [`ExtractSetConfigInputs`](./configwrappers/config.go#L153) function to generate inputs and call [`SetConfig`](./gethwrappers/ManyChainMultiSig.go#L428)
+Once a satisfactory MCMS Membership configuration is constructed, users can use the [`ExtractSetConfigInputs`](config/config.go#L153) function to generate inputs and call [`SetConfig`](./gethwrappers/ManyChainMultiSig.go#L428)
 
 Note: Signers cannot be repeated in this configuration (i.e. they cannot belong to multiple groups)
