+++ pkg/gethwrappers/ManyChainMultiSig.go
