+++ pkg/proposal/mcms/proposal.go
@@ -4,7 +4,7 @@ import (
 	"time"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/smartcontractkit/ccip-owner-contracts/tools/errors"
+	"github.com/smartcontractkit/ccip-owner-contracts/pkg/errors"
 )
 
 type ChainMetadata struct {
