+++ pkg/proposal/timelock/operations.go
@@ -1,6 +1,6 @@
 package timelock
 
-import "github.com/smartcontractkit/ccip-owner-contracts/tools/proposal/mcms"
+import "github.com/smartcontractkit/ccip-owner-contracts/pkg/proposal/mcms"
 
 type BatchChainOperation struct {
 	ChainIdentifier mcms.ChainIdentifier `json:"chainIdentifier"`
